Bulk of the legwork
in this update i have added functionality for almost all of the page. the words are randomly generated with varying amounts of vowels to ensure a word can always be made. the timer is added and functional and the API fetch to confirm the word being given is indeed a real word has been implemented. There are many missing features here still however such as checking the word uses only the letters given on screen. etc.