Finishing out functionality and features
Today i spent the day implementing the missing features to get the app to run such as the buttons functionality and the randomaization of the generated letter sequence. I also added the media queries to make the app responsive. I added a light and dark mode toggle button. I also added some form validation. i am not preventing the user from entering anything unless the round has not started however i am ensuring that they cannot be considered as correct answers. Next i implemented a persistent highscore and as it is only a single number that i need to keep stored i am utilizing local storage for this. I have also added a rules modal to explain the games rules and scoring system.